edit

+++ Notification/script.js
@@ -6,12 +6,17 @@ hamburger_menu.addEventListener("click", () => {
 });
 
 CloseButtons = document.getElementsByClassName('closeButton');
-
-for (button in CloseButtons, noti in notificationList) {
-  button.addEventListener('click', () => {
-    let email = user.email;
-    data = {};
-    data[email] = firebase.firestore.FieldValue.delete();
-    db.collection('events').doc(noti).update(data);
-  })
-}
+await firebase.auth().onAuthStateChanged(async (user) => {
+  if (user) {
+    await getDocumentData(db.collection("users").doc(user.uid));
+    let notificationList = userData['participatedEvents'];
+    for (button in CloseButtons, noti in notificationList) {
+      button.addEventListener('click', () => {
+        let email = user.email;
+        data = {};
+        data[email] = firebase.firestore.FieldValue.delete();
+        db.collection('events').doc(noti).update(data);
+      })
+    };
+  }}
+);
